+++ js/source-data.js
@@ -1,0 +1,71 @@
+const PHOTO_COUNT = 25;
+const COMMENT_COUNT = 50;
+
+const NAMES = [
+  'Алиса',
+  'Александра',
+  'Анжелика',
+  'Анна',
+  'Валентина',
+  'Валерия',
+  'Вера',
+  'Вероника',
+  'Виктория',
+  'Галина',
+  'Дарья',
+  'Диана',
+  'Евгения',
+  'Екатерина',
+  'Алёна',
+  'Елена',
+  'Елизавета',
+  'Жанна',
+  'Карина',
+  'Кристина',
+  'Ксения',
+  'Иван',
+  'Игорь',
+  'Илья',
+  'Кирилл',
+  'Константин',
+  'Лев',
+  'Леонид',
+  'Максим',
+  'Марк',
+  'Матвей',
+  'Михаил',
+  'Никита',
+  'Роман',
+  'Руслан',
+  'Сергей',
+  'Степан',
+  'Тимур',
+  'Фёдор',
+  'Юрий',
+  'Ярослав',
+]
+
+const COMMENTS = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+]
+
+const DESCRIPTIONS = [
+  'Когда радости нет предела.',
+  'Грусть, я тебя не боюсь.',
+  'Любовь в каждом пикселе.',
+  'Фото, заряженное на позитив.',
+  'Улыбаюсь новому дню.',
+  'Я не плачу — это просто дождь.',
+  'Теплые воспоминания в холодное время года.',
+  'В диком восторге от происходящего.',
+  'Знали бы вы, что у меня на уме.',
+  'Где надо, там и где. Угадаете место?',
+  'Что это, если не любовь?',
+]
+
+export { PHOTO_COUNT, COMMENT_COUNT, NAMES, COMMENTS, DESCRIPTIONS }
